added static_cast to check value

+++ tasks/seq/sharapov_g_sobel/func_tests/main.cpp
@@ -38,7 +38,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_1234) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
+    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
   }
 }
 
@@ -74,7 +74,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2451) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
+    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
   }
 }
 
@@ -110,7 +110,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2423) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
+    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
   }
 }
 
@@ -146,7 +146,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2354) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
+    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
   }
 }
 
@@ -182,6 +182,6 @@ TEST(sharapov_g_sobel_seq, Test_Seed_3453) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
+    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
   }
 }
